+++ src/core/utils/telemetry.ts
@@ -1,6 +1,6 @@
 export const telemetry = {
   name: 'react-native-auth0',
-  version: '5.0.1',
+  version: '5.1.0',
 };
 
 export type Telemetry = {