+++ src/exports/classes.ts
@@ -2,5 +2,7 @@ export {
   AuthError,
   CredentialsManagerError,
   WebAuthError,
+  DPoPError,
 } from '../core/models';
-export { default as Auth0, TimeoutError } from '../index';
+export { default as Auth0 } from '../Auth0';
+export { TimeoutError } from '../core/utils/fetchWithTimeout';